requirements.txt
absl-py==2.0.0
asgiref==3.8.1
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==5.0.6
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.49.0
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==3.3.3
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.8
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.23.4
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.18.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2018.9
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
